Remove index from foreign key (#185)
Mutable index `i` is no longer an inherent attribute of a foreign key -- now it is created and encapsulated only in the places that need it.

+++ src/main/scala/trw/dbsubsetter/fktaskqueue/ChronicleQueueFkTaskWriter.scala
@@ -5,7 +5,7 @@ import trw.dbsubsetter.chronicle.ChronicleQueueFunctions
 import trw.dbsubsetter.db.ColumnTypes.ColumnType
 import trw.dbsubsetter.db.ForeignKeyValue
 
-private[fktaskqueue] final class ChronicleQueueFkTaskWriter(fkOrdinal: Short, columnTypes: Seq[ColumnType]) {
+private[fktaskqueue] final class ChronicleQueueFkTaskWriter(fkOrdinal: Int, columnTypes: Seq[ColumnType]) {
   private[this] val valueWriter: (ValueOut, ForeignKeyValue) => Unit = {
     val singleColumnWriteFunctions: Seq[(ValueOut, Any) => WireOut] =
       columnTypes.map(ChronicleQueueFunctions.singleValueWrite)
